Fixes #8488: Generate a custom index page for the manual
