+++ .gitignore
@@ -14,3 +14,4 @@ webhelp*
 html/
 extensions/
 xsl/links.xsl
+xsl/index.html
