+++ xsl/index.html.tpl
@@ -1,0 +1,39 @@
+<div><h2>Welcome to Rudder User Documentation</h2>
+<p><strong>Rudder</strong> is an easy to use, web-driven, role-based
+                   solution for IT Infrastructure Automation and Compliance.</p>
+
+<div class="tip" style="margin-left: 0.5in; margin-right: 0.5in;"><table border="0" summary="Tip"><tr><td rowspan="2" align="center" valign="top" width="25"><img alt="[Tip]" src="common/images/admon-icons/tip.png"></img></td><th align="left">About this documentation</th></tr><tr><td align="left" valign="top">
+
+<p>You can browse this documentation using the left tree, or use the search field in the header bar. It is also embedded in the Rudder server Web interface for offline usage.</p><p>If you find any issue, lacking information,
+or if you have a suggestion, <a href="https://www.rudder-project.org/redmine/projects/rudder/issues/new?issue[category_id]=19">open an issue</a> in our bugtracker.</p>
+
+</td></tr></table></div>
+                   
+<h2>Getting Started</h2>
+<p>If you are new to Rudder you can follow these steps to discover and start using Rudder:</p>
+<ul>
+<li>Read <a href="_introduction.html">introduction</a> to get familiar with Rudder architecture and concepts.</li>
+<li>Follow the <a href="_installation.html">installation</a> section wich explains the <a href="configure_the_network.html">requirements</a> and the installation steps for the different
+                 types of nodes:</li>
+<ul>
+<li>The central <a href="_install_rudder_server.html">server</a> </li>
+<li>The managed <a href="install-agent.html">nodes</a></li>
+</ul>
+<li>Learn how to use the <a href ="_rudder_web_interface.html">Web interface</a>, to <a href="_node_management_2.html">manage nodes</a> and to <a href="_configuration_management_2.html">configure them</a>.</li>
+</ul>
+<h2>Resources</h2>
+<p>Besides this documentation, you can find a lot of other useful information:</p>
+<ul>
+<li>The <a href="http://www.rudder-project.org/changelog-{rudder_version}">Changelog</a> for this version. It describes the changes in each release.</li>
+<li>The <a href="https://www.rudder-project.org/site">Rudder project website</a>, particularly:
+<ul>
+<li>What is an ESR version? How long are Rudder versions supported? All the answers are in the <a href="https://www.rudder-project.org/site/documentation/faq/#versioning">versionning policy</a>.
+</ul>
+<li>You need help or have questions about Rudder? You can:</li>
+<ul>
+<li>Send an email to the user <a href="https://www.rudder-project.org/site/support/forums-lists/">mailing-list</a></li>
+<li>Join the <a href="https://www.rudder-project.org/site/support/chat-mailing-lists/">IRC channel</a></li>
+</ul>
+<li>You need training,  or commercial support? Normation provides <a href="http://www.normation.com/en/rudder/services/">services</a> for Rudder</li>
+<li>Browse the <a href="http://www.rudder-project.org/foswiki/Development/WebHome">developper documentation</a> if you want to contribute or dive into Rudder internals</li>
+</ul></div>
